Moved hasdisability
