+++ webapp/app/forms/flows/lotse_step_chooser.py
@@ -6,8 +6,8 @@ from flask_babel import _
 from app.forms.flows.step_chooser import StepChooser
 from app.forms.steps.lotse.confirmation import StepSummary
 from app.forms.steps.lotse.steuerminderungen import StepVorsorge, StepAussergBela, StepHaushaltsnaheHandwerker, \
-    StepGemeinsamerHaushalt, StepReligion, StepSpenden, StepSelectStmind, StepHasDisability
-from app.forms.steps.lotse.personal_data import StepSteuernummer, StepPersonA, StepPersonB, StepTelephoneNumber
+    StepGemeinsamerHaushalt, StepReligion, StepSpenden, StepSelectStmind
+from app.forms.steps.lotse.personal_data import StepSteuernummer, StepPersonA, StepPersonB, StepTelephoneNumber, StepHasDisability
 
 _LOTSE_DATA_KEY = 'form_data'
