+++ webapp/app/forms/steps/lotse/personal_data.py
@@ -19,7 +19,7 @@ from app.forms.validations.date_validations import ValidDateOfBirth
 from app.forms.validators import DecimalOnly, IntegerLength, ValidHessenTaxNumber, ValidTaxNumber, ValidTaxNumberLength, \
     ValidIdNr, MaximumLength
 from app.forms.validators import DecimalOnly, IntegerLength
-from app.model.components import TaxNumberStepFormProps, TelephoneNumberProps
+from app.model.components import TaxNumberStepFormProps, TelephoneNumberProps, HasDisabilityProps
 from app.model.components.helpers import form_fields_dict
 from app.model.form_data import show_person_b, FamilienstandModel, JointTaxesModel
 
@@ -436,6 +436,48 @@ class StepTelephoneNumber(LotseFormSteuerlotseStep):
 
         return render_template('react_component.html',
                                component='TelephoneNumberPage',
+                               props=props_dict,
+                               form=self.render_info.form,
+                               header_title=_('form.lotse.header-title'))
+
+
+class StepHasDisability(LotseFormSteuerlotseStep):
+    name = 'has_disability'
+    # TODO: Sanny
+    label = _l('form.lotse.has_disability.label')
+    section_link = SectionLink('mandatory_data', StepFamilienstand.name, _l('form.lotse.mandatory_data.label'))
+
+    class InputForm(SteuerlotseBaseForm):
+        has_disability = RadioField(
+            label="",
+            render_kw={'data_label': _l('form.lotse.has-disability.data_label}')},
+            choices=[('married', _l('form.eligibility.marital_status.married')),
+                     ('single', _l('form.eligibility.marital_status.single')),
+                     ('divorced', _l('form.eligibility.marital_status.divorced')),
+                     ('widowed', _l('form.eligibility.marital_status.widowed')),
+                     ],
+            validators=[InputRequired(_l('validate.input-required'))])
+
+    @classmethod
+    def get_label(cls, data):
+        return cls.label
+
+    def render(self):
+        props_dict = HasDisabilityProps(            
+            step_header={
+                'title': _l('form.lotse.has_disability.title'),
+            },
+            form={
+                'action': self.render_info.submit_url,
+                'csrf_token': generate_csrf(),
+                'show_overview_button': bool(self.render_info.overview_url),
+            },
+            fields=form_fields_dict(self.render_info.form),
+            prev_url=self.render_info.prev_url,
+        ).camelized_dict()
+
+        return render_template('react_component.html',
+                               component='HasDisabilityPage',
                                props=props_dict,
                                form=self.render_info.form,
                                header_title=_('form.lotse.header-title'))
